fix(meta-mcp): address CodeRabbit round 2 findings
